# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodHub API",
        "description": "Multi-channel SaaS platform for restaurants - API Documentation",
        "contact": {
            "email": "admin@foodhub.uz"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "FoodHub API Server"
        }
    ],
    "paths": {
...
        "/api/v1/menus": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get list of menus",
                "operationId": "4512a8cf3e47b6bbb48efa21ad1005e0",
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "query",
                        "description": "Filter by restaurant ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "description": "Filter by channel",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile",
                                "telegram",
                                "whatsapp",
                                "phone",
                                "pos"
                            ],
                            "example": "web"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by menu type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "main",
                                "breakfast",
                                "lunch",
                                "dinner",
                                "drinks",
                                "desserts",
                                "seasonal"
                            ],
                            "example": "main"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menus retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Menus retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "object",
                                                                "example": {
                                                                    "en": "Main Menu",
                                                                    "ru": "Основное меню",
                                                                    "uz": "Asosiy menyu"
                                                                }
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "main"
                                                            },
                                                            "channels": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "restaurant": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menus"
                ],
                "summary": "Create a new menu",
                "operationId": "70898b64abbfbbd24bb791f350e431f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "restaurant_id",
                                    "name",
                                    "type",
                                    "channels"
                                ],
                                "properties": {
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "object",
                                        "example": {
                                            "en": "Main Menu",
                                            "ru": "Основное меню",
                                            "uz": "Asosiy menyu"
                                        }
                                    },
                                    "description": {
                                        "type": "object",
                                        "example": {
                                            "en": "Our main menu",
                                            "ru": "Наше основное меню",
                                            "uz": "Bizning asosiy menyumiz"
                                        }
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "main",
                                            "breakfast",
                                            "lunch",
                                            "dinner",
                                            "drinks",
                                            "desserts",
                                            "seasonal"
                                        ],
                                        "example": "main"
                                    },
                                    "channels": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "web",
                                                "mobile",
                                                "telegram",
                                                "whatsapp",
                                                "phone",
                                                "pos"
                                            ]
                                        }
                                    },
                                    "availability_hours": {
                                        "type": "object"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Menu created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/restaurants/{restaurant_slug}/menus": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get menus for specific restaurant",
                "operationId": "59c427eaac53a2be4fe7bb0a7d4889f6",
                "parameters": [
                    {
                        "name": "restaurant_slug",
                        "in": "path",
                        "description": "Restaurant slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pizza-palace"
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "description": "Filter by channel",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile",
                                "telegram",
                                "whatsapp",
                                "phone",
                                "pos"
                            ],
                            "example": "web"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant menus retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurant menus retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "restaurant": {
                                                    "type": "object"
                                                },
                                                "menus": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "object"
                                                            },
                                                            "products": {
                                                                "type": "array",
                                                                "items": {}
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get menu details with products",
                "operationId": "8c0d54161361f4fe73878816aaf255fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "description": "Filter products by channel",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile",
                                "telegram",
                                "whatsapp",
                                "phone",
                                "pos"
                            ],
                            "example": "web"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Menu retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "object"
                                                },
                                                "description": {
                                                    "type": "object"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "main"
                                                },
                                                "restaurant": {
                                                    "type": "object"
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {your-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}